ros-from-csv/test: strip text representation of b'<string>'
The intermingling of Python3 in this code, and Python2-based code in ROS
Melodic has, at some point, resulted in a binary string, represented in
Python3 as b'<string', to be returned with the whole thing as a string, i.e.
"b'string'".

This is a fairly temporary situation, until support for ROS Melodic is
dropped (it's EOL in May 2023), and shouldn't affect later releases.