web/feeds: allow forms to contain dropdown fields
Form fields that have array values are usually flattend to "paths", that
is, the config

```
"my-widget": {
  "type": "text",
  "form": {
    "fields": {
      "choice": [
        "option-a",
        "option-b",
        "option-c"
      ]
    }
  }
}
```

will produce a widget with three text inputs that look like:
```
my-widget
choice[0]: option-a
choice[1]: option-b
choice[2]: option-c
```

With this change, you can now specify that an array field should be
treated as a `<select>` dropdown by using the `"dropdowns":
["field-name"]` syntax.

For the above example, this would look like

```
"my-widget": {
  "type": "text",
  "form": {
    "fields": {
      "choice": [
        "option-a",
        "option-b",
        "option-c"
      ]
    }
    "dropdowns": ["choice"]
  }
}
```

and produce the widget

```
my-widget
choice: option-a
```

where "option-a" is in a dropdown box that can be changed to "option-b"
or "option-c", and on submit will be sent to the backend as
`choice=option-a`.

In this change, the "Clear" button behavior is also modified so that it
will reset dropdowns to their default value, and changing the value of a
dropdown will re-enable the submit button if it had been previously
disabled by clicking "Clear".