web/feeds: adjust config for showing buttons
Previously, the `text` widget would only show buttons when configured as:
```
"form": {
  "buttons": {
    "show": "true"  <-- NOTE: quoted "true"
  }
}
```

Now, we accept both a boolean `true` as well as an array of button
names. For the `text` widget, the available button names are `["ok",
"clear"]`, whilst for the `start_stop` widget, the button names can be a
combination of `["start", "stop", "clear"]`.

So, to configure a `text` widget with only an `ok` button:
```
"form": {
  "buttons": {
    "show": ["ok"]
  }
}
```

NOTE: This remains backwards-compatible with existing configs.