robosense: increase default --scan-max-missing-packets from 10 to 100
This change is required because of jitter in system packet timestamping.
Whilst the expected packet interval (in dual-return mode) is 667us, and
--scan-max-missing-packets=10 results in a 12,000us threshold, we see
inter-packet gaps that exceed this threshold around 5% of the time. This
results in packets being incorrectly assigned to a new scan, rather than
appended to the current scan.

Empirically, increasing --scan-max-missing-packets to 100 (120,000us) is
sufficient to solve the problem.

This is more a workaround than a fix, a better solution would be to look
at the timestamps provided by the lidar to determine if any packets are
missing.