advanced-navigation/device.cpp: read ahead by one header rather than two
The header is five bytes long but some payloads, e.g. Acknowledge at four
bytes, are smaller than a header.

As such, if we try to read two header's worth of data (10 bytes) and we're
not a busy stream (which is the case if we are just querying or configuring
the device) we'll get a situation where we ask for 10 bytes but there are
only 9 in the buffer and the read will block, probably forever if there's no
subsequent traffic.

Ultimately there's no need to read more than one header's worth of data, so
we'll just do that.