advanced-navigation: handle sent messages that have no response
Generally when a packet is sent to the device there is a response, either
another message (in the case of Request packet), or an Acknowledgement (in
the case of any kind of configuration messages)

However some packet writes do not receive a response. Of the currently
support packets, this is only External Time, but probably would include all
the writeable packets in the State Packets block.