vimba-cat: fix bug in output by switching to c-style stdout.
vimba-cat's usage of serialization::write() appears to suffer from a
thread-related race condition.

Expected streaming output:
```
<header_1><image_1>
<header_2><image_2>
...
```

Faulty output:
```
<header_1><image_1>
<header_2><some of image_2><header_2><rest of image_2>
...
```

I don't understand the actual mechanism at play. It appears related to
the fact that images are written to `std::cout` on a background thread
during the frame callback, and somehow something on the main thread is
causing the header to `flush()`ed twice(?). `-fsanitize=thread` warns
about race conditions to do with `std::cout` but not enough to solve the
problem.

This problem had been discovered before in the context of `Pika XC2`
cameras (see comment in `imaging/cv_mat/serialization.cpp` line 241),
and can be resolved by switching to c-style `write()` -- so that's what
we do in this change.