advanced-navigation-to-csv: remove --wait-for-all option for "all" packet
the receipt of packets is (according to the documentation) very deterministic
so there isn't really a use case where we would want to output only on
receipt of some subset of packets.

Furthermore, the old implementation would always timestamp incorrectly, as it
would output on receipt of the system-state packet, but that would arrive
before the other packets on the same tick. Accordingly it would output with
the data for the other packets from the previous tick.

Instead we output on receipt of the last packet for a given tick. The packet
order is guarenteed.

See section 13.6 (Packet Timing) of the Reference Manual for details.